Move MockAuthServer into utils.
